examples/python-example/requirements.txt
aws-cdk.core==1.81.0
aws-cdk.aws-s3-deployment==1.81.0
aws-cdk.aws-eks==1.81.0
awscdk-81-patch